refactor both restart services into one
