+++ admin/services/data-access/pupil-restart.data.service.js
@@ -8,64 +8,6 @@ const pupilRestartDataService = {}
 /** SQL METHODS **/
 
 /**
- * Find latest restart for pupil
- * @param pupilId
- * @return {Promise.<object>}
- */
-pupilRestartDataService.sqlFindLatestRestart = async function (pupilId) {
-  const sql = `SELECT TOP 1 *
-  FROM [mtc_admin].[pupilRestart]
-  WHERE pupil_id=@pupilId AND isDeleted=0
-  ORDER BY createdAt DESC`
-  const params = [
-    {
-      name: 'pupilId',
-      value: pupilId,
-      type: TYPES.Int
-    }
-  ]
-  const result = await sqlService.query(sql, params)
-  return R.head(result)
-}
-
-/**
- * Find pupil's restart codes
- * @return {Promise.<{id:number, code:string, description: string}[]>}
- */
-pupilRestartDataService.sqlFindRestartCodes = async function () {
-  const sql = `
-      SELECT id, code, description
-        FROM [mtc_admin].[pupilRestartCode]
-       ORDER BY description ASC`
-  return sqlService.query(sql)
-}
-
-/**
- * Find pupil's restart reason description by id
- * @param id
- * @return {Promise.<string>}
- */
-pupilRestartDataService.sqlFindRestartReasonDescById = async function (id) {
-  const sql = `
-  SELECT
-    description
-  FROM [mtc_admin].[restartReasonLookup]
-  WHERE id=@id
-  ORDER BY description ASC`
-  const params = [
-    {
-      name: 'id',
-      value: id,
-      type: TYPES.Int
-    }
-  ]
-  const result = await sqlService.query(sql, params)
-  const obj = R.head(result)
-  // @ts-ignore
-  return R.prop('description', obj)
-}
-
-/**
  * Find restart reasons
  * @return {Promise<any>}
  */
@@ -100,8 +42,7 @@ pupilRestartDataService.sqlMarkRestartAsDeleted = async (restartId, userId) => {
       type: TYPES.Int
     }
   ]
-  const sql = `DECLARE @newCheckId INT;
-               DECLARE @newCheckPupilLoginDate DATETIMEOFFSET;
+  const sql = `DECLARE @newCheckPupilLoginDate DATETIMEOFFSET;
                DECLARE @newCheckReceived BIT;
                DECLARE @newCheckComplete BIT;
                DECLARE @originCheckId INT;
@@ -111,10 +52,9 @@ pupilRestartDataService.sqlMarkRestartAsDeleted = async (restartId, userId) => {
 
                -- Populate the main variables
                SELECT
-                 @newCheckId = check_id,
                  @pupilId = pupil_id,
                  @isDeleted = isDeleted,
-                 @originCheckId = originCheck_id
+                 @originCheckId = check_id
                FROM
                 [mtc_admin].[pupilRestart]
                WHERE id = @restartId;
